[UBLE-11] rename: 파일 위치 변경
