+++ src/main/java/com/ureca/uble/domain/usageHistory/repository/CustomUsageHistoryRepository.java
@@ -1,7 +1,7 @@
 package com.ureca.uble.domain.usageHistory.repository;
 
 import com.ureca.uble.domain.usageHistory.dto.response.UsageHistoryRes;
-import com.ureca.uble.global.dto.response.CursorPageRes;
+import com.ureca.uble.global.response.CursorPageRes;
 
 public interface CustomUsageHistoryRepository {
 	CursorPageRes<UsageHistoryRes> findUsagesByUserId(Long userId, Long lastHistoryId, int size);
