+++ src/main/java/com/ureca/uble/domain/usageHistory/service/UsageHistoryService.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Service;
 
 import com.ureca.uble.domain.usageHistory.dto.response.UsageHistoryRes;
 import com.ureca.uble.domain.usageHistory.repository.UsageHistoryRepository;
-import com.ureca.uble.global.dto.response.CursorPageRes;
+import com.ureca.uble.global.response.CursorPageRes;
 
 import lombok.RequiredArgsConstructor;
